refactor(sync): split i18n-pack-store into store modules behind a facade
